+++ example/lib/main.dart
@@ -27,7 +27,11 @@ final appLinks = AppLinks();
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  betterAuthClient = BetterAuthClient(tokenStore: SharedPreferencesTokenStore(), baseUrl: "http://localhost:3000/api/auth", scheme: "bac://");
+  betterAuthClient = BetterAuthClient(
+    tokenStore: SharedPreferencesTokenStore(),
+    baseUrl: "http://localhost:3000/api/auth",
+    scheme: "bac://",
+  );
   runApp(const MainApp());
 }
 
@@ -110,11 +114,22 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver {
           child: ElevatedButton(
             onPressed: () async {
               try {
-                final res = await betterAuthClient.signIn.social(provider: "github", callbackURL: "/details", errorCallbackURL: "/error");
+                final res = await betterAuthClient.signIn.social(
+                  provider: "github",
+                  callbackURL: "/details",
+                  errorCallbackURL: "/error",
+                );
                 final url = Uri.parse(res.url);
-                final result = await FlutterWebAuth2.authenticate(url: url.toString(), callbackUrlScheme: "bac", options: FlutterWebAuth2Options());
+                final result = await FlutterWebAuth2.authenticate(
+                  url: url.toString(),
+                  callbackUrlScheme: "bac",
+                  options: FlutterWebAuth2Options(),
+                );
                 final resultUrl = Uri.parse(result);
-                await launchUrl(resultUrl, mode: LaunchMode.externalApplication);
+                await launchUrl(
+                  resultUrl,
+                  mode: LaunchMode.externalApplication,
+                );
               } catch (e) {
                 print(e);
               }