+++ lib/models/response/base_response.g.dart
@@ -6,29 +6,16 @@ part of 'base_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-BaseResponse<T> _$BaseResponseFromJson<T>(
-  Map<String, dynamic> json,
-  T Function(Object? json) fromJsonT,
-) => BaseResponse<T>(
-  status: json['status'] as bool,
-  message: _$nullableGenericFromJson(json['message'], fromJsonT),
-);
+BaseResponse<T> _$BaseResponseFromJson<T>(Map<String, dynamic> json, T Function(Object? json) fromJsonT) =>
+    BaseResponse<T>(status: json['status'] as bool, message: _$nullableGenericFromJson(json['message'], fromJsonT));
 
-Map<String, dynamic> _$BaseResponseToJson<T>(
-  BaseResponse<T> instance,
-  Object? Function(T value) toJsonT,
-) => <String, dynamic>{
-  'status': instance.status,
-  if (_$nullableGenericToJson(instance.message, toJsonT) case final value?)
-    'message': value,
-};
+Map<String, dynamic> _$BaseResponseToJson<T>(BaseResponse<T> instance, Object? Function(T value) toJsonT) =>
+    <String, dynamic>{
+      'status': instance.status,
+      if (_$nullableGenericToJson(instance.message, toJsonT) case final value?) 'message': value,
+    };
 
-T? _$nullableGenericFromJson<T>(
-  Object? input,
-  T Function(Object? json) fromJson,
-) => input == null ? null : fromJson(input);
+T? _$nullableGenericFromJson<T>(Object? input, T Function(Object? json) fromJson) =>
+    input == null ? null : fromJson(input);
 
-Object? _$nullableGenericToJson<T>(
-  T? input,
-  Object? Function(T value) toJson,
-) => input == null ? null : toJson(input);
+Object? _$nullableGenericToJson<T>(T? input, Object? Function(T value) toJson) => input == null ? null : toJson(input);