+++ lib/models/response/two_factor/two_factor_totp_uri.g.dart
@@ -9,5 +9,6 @@ part of 'two_factor_totp_uri.dart';
 TwoFactorTOTPURI _$TwoFactorTOTPURIFromJson(Map<String, dynamic> json) =>
     TwoFactorTOTPURI(totpURI: json['totpURI'] as String);
 
-Map<String, dynamic> _$TwoFactorTOTPURIToJson(TwoFactorTOTPURI instance) =>
-    <String, dynamic>{'totpURI': instance.totpURI};
+Map<String, dynamic> _$TwoFactorTOTPURIToJson(TwoFactorTOTPURI instance) => <String, dynamic>{
+  'totpURI': instance.totpURI,
+};