+++ lib/models/response/user_and_token_response.g.dart
@@ -6,13 +6,10 @@ part of 'user_and_token_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-UserAndTokenResponse _$UserAndTokenResponseFromJson(
-  Map<String, dynamic> json,
-) => UserAndTokenResponse(
-  user: User.fromJson(json['user'] as Map<String, dynamic>),
-  token: json['token'] as String,
-);
+UserAndTokenResponse _$UserAndTokenResponseFromJson(Map<String, dynamic> json) =>
+    UserAndTokenResponse(user: User.fromJson(json['user'] as Map<String, dynamic>), token: json['token'] as String);
 
-Map<String, dynamic> _$UserAndTokenResponseToJson(
-  UserAndTokenResponse instance,
-) => <String, dynamic>{'user': instance.user, 'token': instance.token};
+Map<String, dynamic> _$UserAndTokenResponseToJson(UserAndTokenResponse instance) => <String, dynamic>{
+  'user': instance.user,
+  'token': instance.token,
+};